move reference guide helpers
