+++ apps/src/util/referenceGuideHelpers.js
